requirements.txt
rpaframework==9.6.0
google-api-python-client==2.23.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
ta-bitwarden-cli==0.6.2